Breaking change in code structure
Now tasks states are managed in one place to avoid typos.
Also was developed the inbox screen and stablish Redux as the
store for the app.
In Storybook was mocked the Redux store and was divided the
original store into slice and reducer in order to manage
just one place of reducers instead of create a new one
each time we need a new mocked store.